module2-llm/requirements.txt
requests>=2.25.0
numpy>=1.19.0
python-dotenv>=0.15.0
sentence-transformers>=2.2.0
scikit-learn>=1.0.0  # For clustering in vector_store.py
